[WIP] Add KFP 2.0 compatibility on SDK and API levels
