+++ mlrun/pipelines/common/helpers.py
@@ -1,10 +1,29 @@
+# Copyright 2024 Iguazio
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 import inspect
-from collections.abc import MutableMapping
-from typing import Any, Dict, Iterator, NoReturn
+from collections.abc import Iterator, MutableMapping
+from typing import Any, NoReturn
+
+project_annotation = "mlrun/project"
+run_annotation = "mlrun/pipeline-step-type"
+function_annotation = "mlrun/function-uri"
 
 
 class FlexibleMapper(MutableMapping):
-    _external_data: Dict
+    _external_data: dict
 
     def __init__(self, external_data: Any):
         if isinstance(external_data, dict):
@@ -12,7 +31,7 @@ class FlexibleMapper(MutableMapping):
         elif hasattr(external_data, "to_dict"):
             self._external_data = external_data.to_dict()
 
-    # TODO: decide if we should kill the dict compatibility layer on get, set and del while using dict syntax
+    # TODO: decide if we should kill the dict compatibility layer on get, set and del
     def __getitem__(self, key: str) -> Any:
         try:
             return getattr(self, key)
@@ -45,10 +64,5 @@ class FlexibleMapper(MutableMapping):
     def __bool__(self) -> bool:
         return bool(self._external_data)
 
-    def to_dict(self) -> Dict:
+    def to_dict(self) -> dict:
         return {k: v for k, v in self}
-
-
-project_annotation = "mlrun/project"
-run_annotation = "mlrun/pipeline-step-type"
-function_annotation = "mlrun/function-uri"
