+++ mlrun/pipelines/common/ops.py
@@ -1,14 +1,347 @@
-import inflection
+# Copyright 2024 Iguazio
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+import json
+import os
+from typing import Union
 
 import mlrun
+from mlrun.config import config
+from mlrun.errors import err_to_str
+from mlrun.model import HyperParamOptions, RunSpec
+from mlrun.utils import get_in, get_workflow_url, is_ipython, logger, version
 
-from mlrun.pipelines.common.helpers import project_annotation, run_annotation, function_annotation
-from mlrun.utils import get_in, logger
+# default KFP artifacts and output (ui metadata, metrics etc.)
+# directories to /tmp to allow running with security context
+KFPMETA_DIR = os.environ.get("KFPMETA_OUT_DIR", "/tmp")
+KFP_ARTIFACTS_DIR = os.environ.get("KFP_ARTIFACTS_DIR", "/tmp")
+
+
+class PipelineRunType:
+    run = "run"
+    build = "build"
+    deploy = "deploy"
+
+
+def mlrun_op(
+    name: str = "",
+    project: str = "",
+    function=None,
+    func_url=None,
+    image: str = "",
+    runobj=None,
+    command: str = "",
+    secrets: list = None,
+    params: dict = None,
+    job_image=None,
+    hyperparams: dict = None,
+    param_file: str = "",
+    labels: dict = None,
+    selector: str = "",
+    inputs: dict = None,
+    outputs: list = None,
+    in_path: str = "",
+    out_path: str = "",
+    rundb: str = "",
+    mode: str = "",
+    handler: str = "",
+    more_args: list = None,
+    hyper_param_options=None,
+    verbose=None,
+    scrape_metrics=False,
+    returns: list[Union[str, dict[str, str]]] = None,
+    auto_build: bool = False,
+):
+    """mlrun KubeFlow pipelines operator, use to form pipeline steps
+
+    when using kubeflow pipelines, each step is wrapped in an mlrun_op
+    one step can pass state and data to the next step, see example below.
+
+    :param name:    name used for the step
+    :param project: optional, project name
+    :param image:   optional, run container image (will be executing the step)
+                    the container should host all required packages + code
+                    for the run, alternatively user can mount packages/code via
+                    shared file volumes like v3io (see example below)
+    :param function: optional, function object
+    :param func_url: optional, function object url
+    :param command: exec command (or URL for functions)
+    :param secrets: extra secrets specs, will be injected into the runtime
+                    e.g. ['file=<filename>', 'env=ENV_KEY1,ENV_KEY2']
+    :param params:  dictionary of run parameters and values
+    :param hyperparams: dictionary of hyper parameters and list values, each
+                        hyperparam holds a list of values, the run will be
+                        executed for every parameter combination (GridSearch)
+    :param param_file:  a csv/json file with parameter combinations, first csv row hold
+                        the parameter names, following rows hold param values
+    :param selector: selection criteria for hyperparams e.g. "max.accuracy"
+    :param hyper_param_options: hyper param options class, see: :py:class:`~mlrun.model.HyperParamOptions`
+    :param labels:   labels to tag the job/run with ({key:val, ..})
+    :param inputs:   dictionary of input objects + optional paths (if path is
+                     omitted the path will be the in_path/key.
+    :param outputs:  dictionary of output objects + optional paths (if path is
+                     omitted the path will be the out_path/key.
+    :param in_path:  default input path/url (prefix) for inputs
+    :param out_path: default output path/url (prefix) for artifacts
+    :param rundb:    path for rundb (or use 'MLRUN_DBPATH' env instead)
+    :param mode:     run mode, e.g. 'pass' for using the command without mlrun wrapper
+    :param handler   code entry-point/handler name
+    :param job_image name of the image user for the job
+    :param verbose:  add verbose prints/logs
+    :param scrape_metrics:  whether to add the `mlrun/scrape-metrics` label to this run's resources
+    :param returns: List of configurations for how to log the returning values from the handler's run (as artifacts or
+                    results). The list's length must be equal to the amount of returning objects. A configuration may be
+                    given as:
+
+                    * A string of the key to use to log the returning value as result or as an artifact. To specify
+                      The artifact type, it is possible to pass a string in the following structure:
+                      "<key> : <type>". Available artifact types can be seen in `mlrun.ArtifactType`. If no artifact
+                      type is specified, the object's default artifact type will be used.
+                    * A dictionary of configurations to use when logging. Further info per object type and artifact
+                      type can be given there. The artifact key must appear in the dictionary as "key": "the_key".
+    :param auto_build: when set to True and the function require build it will be built on the first
+                       function run, use only if you dont plan on changing the build config between runs
+
+    :returns: KFP step operation
+
+    Example:
+    from kfp import dsl
+    from mlrun import mlrun_op
+    from mlrun.platforms import mount_v3io
+
+    def mlrun_train(p1, p2):
+    return mlrun_op('training',
+                    command = '/User/kubeflow/training.py',
+                    params = {'p1':p1, 'p2':p2},
+                    outputs = {'model.txt':'', 'dataset.csv':''},
+                    out_path ='v3io:///projects/my-proj/mlrun/{{workflow.uid}}/',
+                    rundb = '/User/kubeflow')
+
+    # use data from the first step
+    def mlrun_validate(modelfile):
+        return mlrun_op('validation',
+                    command = '/User/kubeflow/validation.py',
+                    inputs = {'model.txt':modelfile},
+                    out_path ='v3io:///projects/my-proj/{{workflow.uid}}/',
+                    rundb = '/User/kubeflow')
+
+    @dsl.pipeline(
+        name='My MLRUN pipeline', description='Shows how to use mlrun.'
+    )
+    def mlrun_pipeline(
+        p1 = 5 , p2 = '"text"'
+    ):
+        # run training, mount_v3io will mount "/User" into the pipeline step
+        train = mlrun_train(p1, p2).apply(mount_v3io())
+
+        # feed 1st step results into the second step
+        validate = mlrun_validate(
+            train.outputs['model-txt']).apply(mount_v3io())
+
+    """
+    # TODO: improve code distribution to avoid circular imports
+    from mlrun.pipelines.ops import generate_pipeline_node
+
+    secrets = [] if secrets is None else secrets
+    params = {} if params is None else params
+    hyperparams = {} if hyperparams is None else hyperparams
+    if hyper_param_options and isinstance(hyper_param_options, dict):
+        hyper_param_options = HyperParamOptions.from_dict(hyper_param_options)
+    inputs = {} if inputs is None else inputs
+    returns = [] if returns is None else returns
+    outputs = [] if outputs is None else outputs
+    labels = {} if labels is None else labels
+
+    rundb = rundb or mlrun.db.get_or_set_dburl()
+    cmd = [
+        "python",
+        "-m",
+        "mlrun",
+        "run",
+        "--kfp",
+        "--from-env",
+        "--workflow",
+        "{{workflow.uid}}",
+    ]
+    file_outputs = {}
+
+    runtime = None
+    code_env = None
+    function_name = ""
+    if function:
+        if not func_url:
+            if function.kind in ["", "local"]:
+                image = image or function.spec.image
+                command = command or function.spec.command
+                more_args = more_args or function.spec.args
+                mode = mode or function.spec.mode
+                rundb = rundb or function.spec.rundb
+                code_env = str(function.spec.build.functionSourceCode)
+            else:
+                runtime = str(function.to_dict())
+
+        function_name = function.metadata.name
+        if function.kind == "dask":
+            image = image or function.spec.kfp_image or config.dask_kfp_image
+
+    image = image or config.kfp_image
+
+    if runobj:
+        handler = handler or runobj.spec.handler_name
+        params = params or runobj.spec.parameters
+        hyperparams = hyperparams or runobj.spec.hyperparams
+        param_file = (
+            param_file
+            or runobj.spec.param_file
+            or runobj.spec.hyper_param_options.param_file
+        )
+        hyper_param_options = hyper_param_options or runobj.spec.hyper_param_options
+        selector = (
+            selector or runobj.spec.selector or runobj.spec.hyper_param_options.selector
+        )
+        inputs = inputs or runobj.spec.inputs
+        returns = returns or runobj.spec.returns
+        outputs = outputs or runobj.spec.outputs
+        in_path = in_path or runobj.spec.input_path
+        out_path = out_path or runobj.spec.output_path
+        secrets = secrets or runobj.spec.secret_sources
+        project = project or runobj.metadata.project
+        labels = runobj.metadata.labels or labels
+        verbose = verbose or runobj.spec.verbose
+        scrape_metrics = scrape_metrics or runobj.spec.scrape_metrics
+
+    outputs = RunSpec.join_outputs_and_returns(outputs=outputs, returns=returns)
+
+    if not name:
+        if not function_name:
+            raise ValueError("name or function object must be specified")
+        name = function_name
+        if handler:
+            short_name = handler
+            for separator in ["#", "::", "."]:
+                # drop paths, module or class name from short name
+                if separator in short_name:
+                    short_name = short_name.split(separator)[-1]
+            name += "-" + short_name
+
+    if hyperparams or param_file:
+        outputs.append("iteration_results")
+    if "run_id" not in outputs:
+        outputs.append("run_id")
+
+    params = params or {}
+    hyperparams = hyperparams or {}
+    inputs = inputs or {}
+    returns = returns or []
+    secrets = secrets or []
+
+    mlrun.runtimes.utils.enrich_run_labels(labels)
+
+    if name:
+        cmd += ["--name", name]
+    if func_url:
+        cmd += ["-f", func_url]
+    for secret in secrets:
+        cmd += ["-s", f"{secret['kind']}={secret['source']}"]
+    for param, val in params.items():
+        cmd += ["-p", f"{param}={val}"]
+    for xpram, val in hyperparams.items():
+        cmd += ["-x", f"{xpram}={val}"]
+    for input_param, val in inputs.items():
+        cmd += ["-i", f"{input_param}={val}"]
+    for log_hint in returns:
+        cmd += [
+            "--returns",
+            json.dumps(log_hint) if isinstance(log_hint, dict) else log_hint,
+        ]
+    for label, val in labels.items():
+        cmd += ["--label", f"{label}={val}"]
+    for output in outputs:
+        cmd += ["-o", str(output)]
+        file_outputs[
+            output.replace(".", "_")
+        ] = f"/tmp/{output}"  # not using path.join to avoid windows "\"
+    if project:
+        cmd += ["--project", project]
+    if handler:
+        cmd += ["--handler", handler]
+    if runtime:
+        cmd += ["--runtime", runtime]
+    if in_path:
+        cmd += ["--in-path", in_path]
+    if out_path:
+        cmd += ["--out-path", out_path]
+    if param_file:
+        cmd += ["--param-file", param_file]
+    if hyper_param_options:
+        cmd += ["--hyper-param-options", hyper_param_options.to_json()]
+    if selector:
+        cmd += ["--selector", selector]
+    if job_image:
+        cmd += ["--image", job_image]
+    if mode:
+        cmd += ["--mode", mode]
+    if verbose:
+        cmd += ["--verbose"]
+    if scrape_metrics:
+        cmd += ["--scrape-metrics"]
+    if auto_build:
+        cmd += ["--auto-build"]
+    if more_args:
+        cmd += more_args
+
+    registry = get_default_reg()
+    if image and image.startswith("."):
+        if registry:
+            image = f"{registry}/{image[1:]}"
+        else:
+            raise ValueError("local image registry env not found")
+
+    image = mlrun.utils.enrich_image_url(
+        image, mlrun.get_version(), str(version.Version().get_python_version())
+    )
+
+    return generate_pipeline_node(
+        project,
+        name,
+        image,
+        cmd + [command],
+        file_outputs,
+        function,
+        func_url,
+        scrape_metrics,
+        code_env,
+        registry,
+    )
+
+
+def get_default_reg():
+    if config.httpdb.builder.docker_registry:
+        return config.httpdb.builder.docker_registry
+    namespace_domain = os.environ.get("IGZ_NAMESPACE_DOMAIN", None)
+    if namespace_domain is not None:
+        return f"docker-registry.{namespace_domain}:80"
+    return ""
 
 
 def format_summary_from_kfp_run(
     kfp_run, project=None, run_db: "mlrun.db.RunDBInterface" = None
 ):
+    # TODO: improve code distribution to avoid circular imports
+    from mlrun.pipelines.ops import generate_kfp_dag_and_resolve_project
+
     override_project = project if project and project != "*" else None
     dag, project, message = generate_kfp_dag_and_resolve_project(
         kfp_run, override_project
@@ -42,47 +375,62 @@ def format_summary_from_kfp_run(
     return short_run
 
 
-def generate_kfp_dag_and_resolve_project(run, project=None):
-    workflow = run.workflow_manifest
-    if not workflow:
-        return None, project, None
+def show_kfp_run(run, clear_output=False):
+    # TODO: review graph rendering functionality with KFP 2.0 data schemas
+    phase_to_color = {
+        mlrun.run.RunStatuses.failed: "red",
+        mlrun.run.RunStatuses.succeeded: "green",
+        mlrun.run.RunStatuses.skipped: "white",
+    }
+    runtype_to_shape = {
+        PipelineRunType.run: "ellipse",
+        PipelineRunType.build: "box",
+        PipelineRunType.deploy: "box3d",
+    }
+    if not run or "graph" not in run:
+        return
+    if is_ipython:
+        try:
+            from graphviz import Digraph
+        except ImportError:
+            return
 
-    templates = {}
-    for template in workflow["spec"]["templates"]:
-        project = project or get_in(
-            template, ["metadata", "annotations", project_annotation], ""
-        )
-        name = template["name"]
-        templates[name] = {
-            "run_type": get_in(
-                template, ["metadata", "annotations", run_annotation], ""
-            ),
-            "function": get_in(
-                template, ["metadata", "annotations", function_annotation], ""
-            ),
-        }
+        try:
+            graph = run["graph"]
+            dag = Digraph("kfp", format="svg")
+            dag.attr(compound="true")
 
-    nodes = workflow["status"].get("nodes", {})
-    dag = {}
-    for node in nodes.values():
-        name = node["displayName"]
-        record = {
-            k: node[k] for k in ["phase", "startedAt", "finishedAt", "type", "id"]
-        }
+            for key, node in graph.items():
+                if node["type"] != "DAG" or node["parent"]:
+                    shape = "ellipse"
+                    if node.get("run_type"):
+                        shape = runtype_to_shape.get(node["run_type"], None)
+                    elif node["phase"] == "Skipped" or (
+                        node["type"] == "DAG" and node["name"].startswith("condition-")
+                    ):
+                        shape = "diamond"
+                    dag.node(
+                        key,
+                        label=node["name"],
+                        fillcolor=phase_to_color.get(node["phase"], None),
+                        style="filled",
+                        shape=shape,
+                        tooltip=node.get("error", None),
+                    )
+                    for child in node.get("children") or []:
+                        dag.edge(key, child)
 
-        # snake case
-        # align kfp fields to mlrun snake case convention
-        # create snake_case for consistency.
-        # retain the camelCase for compatibility
-        for key in list(record.keys()):
-            record[inflection.underscore(key)] = record[key]
+            import IPython
 
-        record["parent"] = node.get("boundaryID", "")
-        record["name"] = name
-        record["children"] = node.get("children", [])
-        if name in templates:
-            record["function"] = templates[name].get("function")
-            record["run_type"] = templates[name].get("run_type")
-        dag[node["id"]] = record
+            if clear_output:
+                IPython.display.clear_output(wait=True)
 
-    return dag, project, workflow["status"].get("message", "")
+            run_id = run["run"]["id"]
+            url = get_workflow_url(run["run"]["project"], run_id)
+            href = f'<a href="{url}" target="_blank"><b>click here</b></a>'
+            html = IPython.display.HTML(
+                f"<div>Pipeline running (id={run_id}), {href} to view the details in MLRun UI</div>"
+            )
+            IPython.display.display(html, dag)
+        except Exception as exc:
+            logger.warning(f"failed to plot graph, {err_to_str(exc)}")
