+++ mlrun/pipelines/kfp/v2_0/helpers.py
@@ -1,9 +1,24 @@
+# Copyright 2024 Iguazio
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
 import typing
 
 
 def new_pipe_metadata(
     artifact_path: str = None,
     cleanup_ttl: int = None,
-    op_transformers: typing.List[typing.Callable] = None,
+    op_transformers: list[typing.Callable] = None,
 ):
-    return
+    raise NotImplementedError
