+++ mlrun/pipelines/kfp/v2_0/iguazio.py
@@ -1,3 +1,19 @@
+# Copyright 2024 Iguazio
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+
 def v3io_cred(api="", user="", access_key=""):
     """
     Modifier function to copy local v3io env vars to container
@@ -7,7 +23,7 @@ def v3io_cred(api="", user="", access_key=""):
         train = train_op(...)
         train.apply(use_v3io_cred())
     """
-    return
+    raise NotImplementedError
 
 
 def mount_v3io(
@@ -27,8 +43,8 @@ def mount_v3io(
     :param secret:          k8s secret name which would be used to get the username and access key to auth against v3io.
     :param volume_mounts:   list of VolumeMount. empty volume mounts & remote will default to mount /v3io & /User.
     """
-    return
+    raise NotImplementedError
 
 
 def mount_v3iod(namespace, v3io_config_configmap):
-    return
+    raise NotImplementedError
