+++ requirements.txt
@@ -4,9 +4,7 @@ GitPython~=3.1, >=3.1.41
 aiohttp~=3.9
 aiohttp-retry~=2.8
 click~=8.1
-# 3.0/3.2 iguazio system uses 1.0.1, but we needed >=1.6.0 to be compatible with k8s>=12.0 to fix scurity issue
-# since the sdk is still mark as beta (and not stable) I'm limiting to only patch changes
-kfp~=1.8
+kfp[kubernetes]~=2.5
 nest-asyncio~=1.0
 ipython~=8.10
 nuclio-jupyter~=0.9.16
