+++ server/api/api/endpoints/pipelines.py
@@ -28,6 +28,7 @@ import server.api.crud
 import server.api.utils.auth.verifier
 import server.api.utils.singletons.k8s
 from mlrun.config import config
+from mlrun.pipelines.models import PipelineManifest
 from mlrun.utils import logger
 from server.api.api import deps
 from server.api.api.utils import log_and_raise
@@ -209,8 +210,6 @@ async def _create_pipeline(
     content_type = request.headers.get("content-type", "")
 
     # otherwise, best effort - try to parse it from the body - if successful - perform auth check - otherwise explode
-    # TODO: this check fails with KFP SDK 2.0, since it doesn't compile to yaml on the client side anymore
-    # TODO: there might be a workaround if we are able to read the new file format
     project = _try_resolve_project_from_body(content_type, data)
     if not project:
         raise mlrun.errors.MLRunInvalidArgumentError(
@@ -260,5 +259,5 @@ def _try_resolve_project_from_body(
         return None
     workflow_manifest = yaml.load(data, Loader=yaml.FullLoader)
     return server.api.crud.Pipelines().resolve_project_from_workflow_manifest(
-        workflow_manifest
+        PipelineManifest(workflow_manifest)
     )
