+++ server/api/crud/pipelines.py
@@ -33,11 +33,13 @@ import mlrun.utils.helpers
 import mlrun.utils.singleton
 import server.api.api.utils
 from mlrun.errors import err_to_str
-from mlrun.pipelines.models import PipelineRun
+from mlrun.pipelines.mixins import PipelineProviderMixin
+from mlrun.pipelines.models import PipelineExperiment, PipelineRun
 from mlrun.utils import logger
 
 
 class Pipelines(
+    PipelineProviderMixin,
     metaclass=mlrun.utils.singleton.Singleton,
 ):
     def list_pipelines(
@@ -255,9 +257,13 @@ class Pipelines(
 
         try:
             kfp_client = self.initialize_kfp_client(namespace)
-            experiment = kfp_client.create_experiment(name=experiment_name)
-            run = kfp_client.run_pipeline(
-                experiment.id, run_name, pipeline_file.name, params=arguments
+            experiment = PipelineExperiment(
+                kfp_client.create_experiment(name=experiment_name)
+            )
+            run = PipelineRun(
+                kfp_client.run_pipeline(
+                    experiment.id, run_name, pipeline_file.name, params=arguments
+                )
             )
         except Exception as exc:
             logger.warning(
@@ -371,55 +377,6 @@ class Pipelines(
 
     def resolve_project_from_pipeline(self, pipeline: PipelineRun):
         return self.resolve_project_from_workflow_manifest(pipeline.workflow_manifest)
-
-    def resolve_project_from_workflow_manifest(self, workflow_manifest):
-        templates = workflow_manifest.get("spec", {}).get("templates", [])
-        for template in templates:
-            project_from_annotation = (
-                template.get("metadata", {})
-                .get("annotations", {})
-                .get(mlrun.pipelines.common.helpers.project_annotation)
-            )
-            if project_from_annotation:
-                return project_from_annotation
-            command = template.get("container", {}).get("command", [])
-            action = None
-            for index, argument in enumerate(command):
-                if argument == "mlrun" and index + 1 < len(command):
-                    action = command[index + 1]
-                    break
-            if action:
-                if action == "deploy":
-                    project = self._resolve_project_from_command(
-                        command,
-                        hyphen_p_is_also_project=True,
-                        has_func_url_flags=True,
-                        has_runtime_flags=False,
-                    )
-                    if project:
-                        return project
-                elif action == "run":
-                    project = self._resolve_project_from_command(
-                        command,
-                        hyphen_p_is_also_project=False,
-                        has_func_url_flags=True,
-                        has_runtime_flags=True,
-                    )
-                    if project:
-                        return project
-                elif action == "build":
-                    project = self._resolve_project_from_command(
-                        command,
-                        hyphen_p_is_also_project=False,
-                        has_func_url_flags=False,
-                        has_runtime_flags=True,
-                    )
-                    if project:
-                        return project
-                else:
-                    raise NotImplementedError(f"Unknown action: {action}")
-
-        return mlrun.mlconf.default_project
 
     @staticmethod
     def _get_experiment_id_from_run(run: dict) -> str:
