+++ tests/platforms/test_other.py
@@ -16,6 +16,7 @@ import deepdiff
 
 import mlrun
 import mlrun.errors
+import mlrun.pipelines.kfp.v1_8.platforms.other
 
 
 def test_mount_configmap():
@@ -87,7 +88,7 @@ def test_mount_s3():
         "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
     )
     function.apply(
-        mlrun.platforms.mount_s3(
+        mlrun.pipelines.kfp.v1_8.platforms.other.mount_s3(
             aws_access_key="xx", aws_secret_key="yy", endpoint_url="a.b"
         )
     )
@@ -101,7 +102,11 @@ def test_mount_s3():
     function = mlrun.new_function(
         "function-name", "function-project", kind=mlrun.runtimes.RuntimeKinds.job
     )
-    function.apply(mlrun.platforms.mount_s3(secret_name="s", endpoint_url="a.b"))
+    function.apply(
+        mlrun.pipelines.kfp.v1_8.platforms.other.mount_s3(
+            secret_name="s", endpoint_url="a.b"
+        )
+    )
     env_dict = {
         var["name"]: var.get("value", var.get("valueFrom")) for var in function.spec.env
     }
